Rename module file name
